pyproject.toml
[tool.poetry]
name = "faaf-paper-code"
version = "0.1.0"
description = ""
authors = ["vasiliskatr <vasileios.katranidis@immo.capital>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
pydantic = "^2.6.4"
openai = "^1.14.2"
anthropic = "^0.21.1"
mistralai = "^0.1.6"
python-dotenv = "^1.0.1"
datasets = "^2.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
